a ordenacao estava certo, porem para confirma isso foi retirando se os elementos do topo, e ele devolvia em ordem correta

+++ src/Utilidades.java
@@ -24,9 +24,9 @@ public class Utilidades {
         return resultados;
     }
 
-    public static double calculaDistancia(double x1, double y1, double x2, double y2){
+    public static int calculaDistancia(double x1, double y1, double x2, double y2){
         double resultado = 0;
         resultado = Math.sqrt( (Math.pow((x2 - x1), 2) + Math.pow((y2 - y1), 2)) );
-        return resultado;
+        return (int) resultado;
     }
 }
